Renamed the file containing the elph_qe_to_phoebe class to match the class name.
