+++ src/apps/app.cpp
@@ -6,7 +6,7 @@
 #include "io.h"
 #include "phonon_transport_app.h"
 #include "electron_wannier_transport_app.h"
-#include "elph_bloch_to_wannier_app.h"
+#include "elph_qe_to_phoebe_app.h"
 #include "polarization_app.h"
 #include "utilities.h"
 #include <cmath>
