+++ test/interactionElPhTest.cpp
@@ -1,4 +1,4 @@
-#include "elph_bloch_to_wannier_app.h"
+#include "elph_qe_to_phoebe_app.h"
 #include "qe_input_parser.h"
 #include "gtest/gtest.h"
